Blogger example: fixes build.
The Blogger example was not building for me when I checked out the repository.
I had to point the project to the new location of the GTLCore.xcodeproj (it was
GTL.xcodeproj).

Likely the other examples will need similar changes.

Since I wasn't sure which files from the BloggerSample.xcodeproj to include, I
also modified .gitignore to match the recommended settings at
https://github.com/github/gitignore/blob/master/Objective-C.gitignore.